исправления по модулям
