+++ js/main.js
@@ -1,6 +1,5 @@
-import {createAnnouncement} from './data.js';
-import {ANNOUNCEMENT_COUNT} from '.constants.js';
+import {createAnnouncement, ANNOUNCEMENT_COUNT} from './data.js';
 
 const announcementArray = Array.from({length:ANNOUNCEMENT_COUNT}, createAnnouncement);
-announcementArray;
+console.log(announcementArray);
